image-captioner/backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
transformers==4.38.1
torch==2.2.1
Pillow==10.2.0
numpy<2
